Stubbed MoodSystem
